service controller impliment

+++ src/tasks/service/tasks.service.ts
@@ -1,29 +1,32 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
-import { InjectRepository } from '@nestjs/typeorm';
+import {Injectable, NotFoundException} from '@nestjs/common';
+import {InjectRepository} from '@nestjs/typeorm';
 import {TaskRepository} from '../repository/task.repository';
-import {Task} from '../entity/task.entity';
 import {GetTasksFilterDto} from '../dto/get-tasks-filter.dto';
+import {User} from '../../auth/entity/user.entity';
+import {Task} from '../entity/task.entity';
 import {CreateTaskDto} from '../dto/create-task.dto';
 import {TaskStatus} from '../enum/task-status.enum';
-
 
 @Injectable()
 export class TasksService {
     constructor(
         @InjectRepository(TaskRepository)
         private taskRepository: TaskRepository,
-    ) {}
-
-    async getAllTasks(): Promise<Task[]> {
-        return this.taskRepository.find();
+    ) {
     }
 
-    getTasksWithFilters(filterDto: GetTasksFilterDto): Promise<Task[]> {
-        return this.taskRepository.getTasksWithFilters(filterDto);
+    async getTasks(
+        filterDto: GetTasksFilterDto,
+        user: User,
+    ): Promise<Task[]> {
+        return this.taskRepository.getTasks(filterDto, user);
     }
 
-    async getTaskById(id: number): Promise<Task> {
-        const found = await this.taskRepository.findOne(id);
+    async getTaskById(
+        id: number,
+        user: User,
+    ): Promise<Task> {
+        const found = await this.taskRepository.findOne({where: {id, userId: user.id}});
 
         if (!found) {
             throw new NotFoundException(`Task with ID "${id}" not found`);
@@ -32,20 +35,30 @@ export class TasksService {
         return found;
     }
 
-    async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-        return this.taskRepository.createTask(createTaskDto);
+    async createTask(
+        createTaskDto: CreateTaskDto,
+        user: User,
+    ): Promise<Task> {
+        return this.taskRepository.createTask(createTaskDto, user);
     }
 
-    async deleteTask(id: number): Promise<void> {
-        const result = await this.taskRepository.delete(id);
+    async deleteTask(
+        id: number,
+        user: User,
+    ): Promise<void> {
+        const result = await this.taskRepository.delete({id, userId: user.id});
 
         if (result.affected === 0) {
             throw new NotFoundException(`Task with ID "${id}" not found`);
         }
     }
 
-    async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-        const task = await this.getTaskById(id);
+    async updateTaskStatus(
+        id: number,
+        status: TaskStatus,
+        user: User,
+    ): Promise<Task> {
+        const task = await this.getTaskById(id, user);
         task.status = status;
         await task.save();
         return task;
